오늘 해야 할것 : 선한쌤과 버그 고치기 , 명희쌤과 버그 고치기 ,  적 최소한 5개 이상 만들기 , 무기 4개 추가, 인트로 씬 내일 해야 할 것 : 사운드 추가 밑 버그 고치기 , 웨이브 추가 나머지 : 애니메이션 고치기, 맵 고치기

+++ Assets/01.Scripts/SO/PlayerDataSO.cs
@@ -8,9 +8,9 @@ public class PlayerDataSO : ScriptableObject
     [Range(5,8)]
     public float Speed; //움직임
     [Range(80,150)]
-    public float Hp; //체력
+    public int Hp; //체력
     [Range(100,150)]
-    public float Stamina; //스태미나
+    public int Stamina; //스태미나
     [Range(30,50)]
     public float Damage; //데미지
     [Range(10,25)]
